pyproject.toml
[project]
name = "unisport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "flask==3.1.0",
    "flask-cors==4.0.2",
    "flask-sqlalchemy==3.1.1",
    "gunicorn==23.0.0",
    "psycopg2==2.9.10",
    "pytest==8.3.5",
    "pytest-cov==2.5.1",
    "sentry-sdk[flask]==2.24.1",
]
